ENH: Les obscurances agafen les dades del volum com a unsigned short.

+++ starviewer/src/core/obscurancethread.h
@@ -45,7 +45,7 @@ public:
 
     /// Assigna l'estimador del gradient, d'on es treuran les normals.
     void setGradientEstimator( vtkEncodedGradientEstimator *gradientEstimator );
-    void setData( const uchar * data, int dataSize, const int dimensions[3], const int increments[3] );
+    void setData( const ushort *data, int dataSize, const int dimensions[3], const int increments[3] );
     void setObscuranceParameters( double obscuranceMaximumDistance, ObscuranceMainThread::Function obscuranceFunction, ObscuranceMainThread::Variant obscuranceVariant, Obscurance *obscurance );
     void setSaliency( const double * saliency, double fxSaliencyA, double fxSaliencyB, double fxSaliencyLow, double fxSaliencyHigh );
     void setPerDirectionParameters( const Vector3 & direction, const Vector3 & forward, const int xyz[3], const int sXYZ[3], const QVector<Vector3> & lineStarts, qptrdiff startDelta );
@@ -75,7 +75,7 @@ private:
     const TransferFunction & m_transferFunction;
     vtkDirectionEncoder * m_directionEncoder;
     const ushort * m_encodedNormals;
-    const uchar * m_data;
+    const ushort *m_data;
     int m_dataSize;
     const int * m_dimensions;
     const int * m_increments;
